feat: library option to automatically empty trash after scan

+++ komga/src/main/kotlin/org/gotson/komga/domain/service/LibraryContentLifecycle.kt
@@ -3,8 +3,10 @@ package org.gotson.komga.domain.service
 import mu.KotlinLogging
 import org.gotson.komga.application.tasks.TaskReceiver
 import org.gotson.komga.domain.model.BookSearch
+import org.gotson.komga.domain.model.DirectoryNotFoundException
 import org.gotson.komga.domain.model.Library
 import org.gotson.komga.domain.model.Media
+import org.gotson.komga.domain.model.ScanResult
 import org.gotson.komga.domain.model.SeriesSearch
 import org.gotson.komga.domain.model.Sidecar
 import org.gotson.komga.domain.persistence.BookRepository
@@ -39,7 +41,12 @@ class LibraryContentLifecycle(
   fun scanRootFolder(library: Library) {
     logger.info { "Updating library: $library" }
     measureTime {
-      val scanResult = fileSystemScanner.scanRootFolder(Paths.get(library.root.toURI()), library.scanForceModifiedTime)
+      val (scanResult, rootFolderInaccessible) = try {
+        fileSystemScanner.scanRootFolder(Paths.get(library.root.toURI()), library.scanForceModifiedTime) to false
+      } catch (e: DirectoryNotFoundException) {
+        ScanResult(emptyMap(), emptyList()) to true
+      }
+
       val scannedSeries =
         scanResult
           .series
@@ -128,40 +135,43 @@ class LibraryContentLifecycle(
         }
       }
 
-      val existingSidecars = sidecarRepository.findAll()
-      scanResult.sidecars.forEach { newSidecar ->
-        val existingSidecar = existingSidecars.firstOrNull { it.url == newSidecar.url }
-        if (existingSidecar == null || existingSidecar.lastModifiedTime.notEquals(newSidecar.lastModifiedTime)) {
-          when (newSidecar.source) {
-            Sidecar.Source.SERIES ->
-              seriesRepository.findByLibraryIdAndUrlOrNull(library.id, newSidecar.parentUrl)?.let { series ->
-                logger.info { "Sidecar changed on disk (${newSidecar.url}, refresh Series for ${newSidecar.type}: $series" }
-                when (newSidecar.type) {
-                  Sidecar.Type.ARTWORK -> taskReceiver.refreshSeriesLocalArtwork(series.id)
+      if (!rootFolderInaccessible) {
+        val existingSidecars = sidecarRepository.findAll()
+        scanResult.sidecars.forEach { newSidecar ->
+          val existingSidecar = existingSidecars.firstOrNull { it.url == newSidecar.url }
+          if (existingSidecar == null || existingSidecar.lastModifiedTime.notEquals(newSidecar.lastModifiedTime)) {
+            when (newSidecar.source) {
+              Sidecar.Source.SERIES ->
+                seriesRepository.findByLibraryIdAndUrlOrNull(library.id, newSidecar.parentUrl)?.let { series ->
+                  logger.info { "Sidecar changed on disk (${newSidecar.url}, refresh Series for ${newSidecar.type}: $series" }
+                  when (newSidecar.type) {
+                    Sidecar.Type.ARTWORK -> taskReceiver.refreshSeriesLocalArtwork(series.id)
+                  }
                 }
-              }
-            Sidecar.Source.BOOK ->
-              bookRepository.findByLibraryIdAndUrlOrNull(library.id, newSidecar.parentUrl)?.let { book ->
-                logger.info { "Sidecar changed on disk (${newSidecar.url}, refresh Book for ${newSidecar.type}: $book" }
-                when (newSidecar.type) {
-                  Sidecar.Type.ARTWORK -> taskReceiver.refreshBookLocalArtwork(book.id)
+              Sidecar.Source.BOOK ->
+                bookRepository.findByLibraryIdAndUrlOrNull(library.id, newSidecar.parentUrl)?.let { book ->
+                  logger.info { "Sidecar changed on disk (${newSidecar.url}, refresh Book for ${newSidecar.type}: $book" }
+                  when (newSidecar.type) {
+                    Sidecar.Type.ARTWORK -> taskReceiver.refreshBookLocalArtwork(book.id)
+                  }
                 }
-              }
+            }
+            sidecarRepository.save(library.id, newSidecar)
           }
-          sidecarRepository.save(library.id, newSidecar)
         }
-      }
 
-      // cleanup sidecars that don't exist anymore
-      scanResult.sidecars.map { it.url }.let { newSidecarsUrls ->
-        existingSidecars
-          .filterNot { existing -> newSidecarsUrls.contains(existing.url) }
-          .let { sidecars ->
-            sidecarRepository.deleteByLibraryIdAndUrls(library.id, sidecars.map { it.url })
-          }
-      }
+        // cleanup sidecars that don't exist anymore
+        scanResult.sidecars.map { it.url }.let { newSidecarsUrls ->
+          existingSidecars
+            .filterNot { existing -> newSidecarsUrls.contains(existing.url) }
+            .let { sidecars ->
+              sidecarRepository.deleteByLibraryIdAndUrls(library.id, sidecars.map { it.url })
+            }
+        }
 
-      cleanupEmptySets()
+        if (library.emptyTrashAfterScan) emptyTrash(library)
+        else cleanupEmptySets()
+      }
     }.also { logger.info { "Library updated in $it" } }
   }
